tried using a double for the wigth

+++ push_swap.h
@@ -1,6 +1,6 @@
 #ifndef PUSH_SWAP_H
 # define PUSH_SWAP_H
-# include "libft.h"
+# include <libft.h>
 # define _FRONT 1
 # define _BACK 0
 
@@ -28,7 +28,7 @@ typedef struct		s_stack
 
 typedef struct		s_node
 {
-	long			weight;
+	double			weight;
 	struct s_node	*parent;
 	t_stack			*a;
 	t_stack			*b;
@@ -119,7 +119,7 @@ void		expand_open_set(t_nodelist **open, t_pack *final_state,
 				t_pack *pack, t_stack *stack_a);
 char		**create_list_of_all_moves(void);
 void		sort_arr(int **arr, int size);
-long		calc_weight(t_node *node, t_pack *final, t_pack *pack);
+double		calc_weight(t_node *node, t_pack *final, t_pack *pack);
 void		print_move_list(t_node *node);
 char		**moves_to_current(t_node *node);
 void		apply_rule(t_stack **a, t_stack **b, char *rule);
